django/requirements/base.txt
Django>=5.2,<5.3
django-environ>=0.12.0

djangorestframework>=3.16.0
django-cors-headers>=4.9.0

djangorestframework-simplejwt>=5.5.0
dj-rest-auth>=7.0.0

argon2-cffi>=25.0.0

playwright>=1.58.0
qrcode[pil]>=8.2

setuptools>=65

Pillow==10.4.0